+++ src/index.ts
@@ -1,22 +1,32 @@
-import { gt } from "./ddzy/utility/others/gt";
+import { capitalize } from "./ddzy/utility/string/capitalize";
 
-// ? true
-const s1 = [10, 5];
-const p1 = gt(s1[0], s1[1]);
+// ?
+const s1 = '';
+const p1 = capitalize(s1);
 console.log(p1);
 
-console.log('-----------------');
+console.log('-------------------');
 
-// ? false
-const s2 = [10, 10];
-const p2 = gt(s2[0], s2[1]);
+// ?
+const s2 = 'ZHAO';
+const p2 = capitalize(s2);
 console.log(p2);
+
+console.log('-------------------');
+
+// ?
+const s3 = 'duan';
+const p3 = capitalize(s3);
+console.log(p3);
 
 console.log('----------------------');
 
-// ? false
-const s3 = [10, 20];
-const p3 = gt(s3[0], s3[1]);
-console.log(p3);
+// ?
+const s4 = '_DDZY';
+const p4 = capitalize(s4);
+console.log(p4);
 
-console.log('-----------------------');
+// ?
+const s5 = '  ___&*duanzhaoyang';
+const p5 = capitalize(s5);
+console.log(p5);